+++ src/str.ts
@@ -275,7 +275,7 @@ export class Str {
    *
    * @returns {Boolean}
    */
-  notEquals (value: string): boolean {
+  notEquals (value: string | Str): boolean {
     return !this.equals(value)
   }
 
@@ -366,7 +366,16 @@ export class Str {
    * @returns {Boolean}
    */
   isLowercase (): boolean {
-    return this.value === this.lower().get()
+    return this.equals(this.lower())
+  }
+
+  /**
+   * Determine whether the string is lowercase and consists of letters only.
+   *
+   * @returns {Boolean}
+   */
+  isLowerLetters (): boolean {
+    return this.equals(this.lower()) && this.notEquals(this.upper())
   }
 
   /**
@@ -424,7 +433,7 @@ export class Str {
    * @returns {Boolean}
    */
   isUppercase (): boolean {
-    return this.value === this.upper().get()
+    return this.equals(this.upper())
   }
 
   /**